Added multidirectional movement, Fixed Truck Bug
This version adds multidimensional movement to accompany the larger map size. It is only temporary! When we get graphics running over it we won't need it. It will only be useful to see the proposed movement and the cost. Also I THINK I've got truck working. After selecting him you'd need to choose NOT move him and then choose to hold a piece and it will get held.